src/requirements.txt
pika
redis
Django
gunicorn
drf-yasg
DateTime
channels
python-dotenv
django-filter
channels-redis
django-jazzmin
psycopg2-binary
uvicorn[standard]
django-cors-headers
djangorestframework